Move filters outside of ResourceBrowser component
